Don't import numba when dascore is imported
maybe_numba_jit imports numba and builds the dispatcher when it decorates,
so every jit'ed kernel pulled numba in at import time. Patch binds the
transform functions as class attributes, which makes dascore.transform (and
therefore taup and kurtosis) part of the eager import graph, so importing
numba cost every user ~190 ms whether or not they ever called tau_p or
kurtosis.

Move the decorated kernels into private _taup_kernels and _kurtosis_kernels
modules and import them from inside tau_p and kurtosis instead. The kernels
are unchanged and maybe_numba_jit still returns a real numba dispatcher, so
kernels calling other kernels in nopython mode keep working.

Median `import dascore` drops from 1082 ms to 896 ms.
